Update SeedMake to add support for new config('seeders.dir') format
The config('seeders.dir') format was previously changed to be an array of paths rather than a single path. However SeedMake wasn't updated to take this into account. Also another change was made which expects the paths in the config to already have database_path(), therefore SeedMake will no longer add this.